+++ app/(main)/(tabs)/(home)/index.tsx
@@ -1,2 +1,1 @@
-export { default } from '@/modules/home/screens/HomeScreen';
-
+export { default } from "@/features/home/screens/HomeScreen";