+++ app/(main)/(tabs)/(profile)/index.tsx
@@ -1,2 +1,1 @@
-export { default } from '@/modules/profile/screens/ProfileScreen';
-
+export { default } from "@/features/profile/screens/ProfileScreen";